- transformed_shape_op() ported to gotoized c w/commented-out inline asm. keep the original code in intermingled comments until the function has been thoroughly tested - various 32bit/sdl fixes

+++ src/restunts/c/memmgr.h
@@ -8,9 +8,9 @@
 #pragma pack (push, 1)
 struct MEMCHUNK {
 	char resname[12];
-	unsigned short ressize;
-	unsigned short resofs;
-	unsigned short resunk;
+	unsigned ressize;
+	unsigned resofs;
+	unsigned resunk;
 };
 #pragma pack (pop)
